+++ jscomp/ml/translcore.ml
@@ -762,7 +762,14 @@ and transl_exp0 (e : Typedtree.expression) : Lambda.lambda =
       let inlined, funct =
         Translattribute.get_and_remove_inlined_attribute funct
       in
-      transl_apply ~inlined (transl_exp funct) oargs e.exp_loc
+      let uncurried_partial_application =
+        let uncurried_partial_app = Ext_list.exists e.exp_attributes (fun ({txt },_) -> txt = "res.partial") in
+        if uncurried_partial_app then
+          let arity_opt = Ast_uncurried.uncurried_type_get_arity_opt ~env:funct.exp_env funct.exp_type in   
+          arity_opt
+        else
+          None in
+      transl_apply ~inlined ~uncurried_partial_application (transl_exp funct) oargs e.exp_loc
   | Texp_match (arg, pat_expr_list, exn_pat_expr_list, partial) ->
       transl_match e arg pat_expr_list exn_pat_expr_list partial
   | Texp_try (body, pat_expr_list) ->
@@ -978,7 +985,7 @@ and transl_cases_try cases =
   in
   List.map transl_case_try cases
 
-and transl_apply ?(inlined = Default_inline) lam sargs loc =
+and transl_apply ?(inlined = Default_inline) ?(uncurried_partial_application=None) lam sargs loc =
   let lapply funct args =
     match funct with
     (* Attention: This may not be what we need to change the application arity*)
@@ -1028,11 +1035,36 @@ and transl_apply ?(inlined = Default_inline) lam sargs loc =
     | (Some arg, optional) :: l -> build_apply lam ((arg, optional) :: args) l
     | [] -> lapply lam (List.rev_map fst args)
   in
-  (build_apply lam []
-     (List.map
-        (fun (l, x) -> (may_map transl_exp x, Btype.is_optional l))
-        sargs)
-    : Lambda.lambda)
+  match uncurried_partial_application with
+    | Some arity when arity > List.length sargs ->
+    let extra_arity = arity - List.length sargs in
+    let none_ids = ref [] in
+    let args = Ext_list.filter_map sargs (function
+     | _, Some e ->
+        Some (transl_exp e)
+     | _, None ->
+        let id_arg = Ident.create "none" in
+        none_ids := id_arg :: !none_ids;
+        Some (Lvar id_arg)) in
+    let extra_ids = ref [] in
+    extra_ids := Ident.create "extra" :: !extra_ids;
+    let extra_ids = Array.init extra_arity (fun _ -> Ident.create "extra") |> Array.to_list in
+    let extra_args = Ext_list.map extra_ids (fun id -> Lvar id) in
+    let ap_args = args @ extra_args in
+    let l0 = Lapply { ap_func = lam; ap_args; ap_inlined = inlined; ap_loc = loc } in
+    Lfunction
+      {
+        params = List.rev_append !none_ids extra_ids ;
+        body = l0;
+        attr = default_function_attribute;
+        loc;
+      }
+  | _ ->
+    (build_apply lam []
+      (List.map
+          (fun (l, x) -> (may_map transl_exp x, Btype.is_optional l))
+          sargs)
+      : Lambda.lambda)
 
 and transl_function loc partial param cases =
   match cases with