+++ jscomp/test/UncurriedAlways.js
@@ -24,8 +24,8 @@ console.log(a);
       return x + 1 | 0;
     });
 
-function ptl(param) {
-  return foo(10, param);
+function ptl(extra) {
+  return 10 + extra | 0;
 }
 
 function foo2(x, y) {
@@ -56,6 +56,108 @@ function inl2(x, y) {
   return x + y | 0;
 }
 
+function foo$1(x, y, z) {
+  return [
+          x,
+          y,
+          z
+        ];
+}
+
+function ptl$1(none, extra) {
+  return [
+          none,
+          "y",
+          extra
+        ];
+}
+
+var a1 = [
+  "x",
+  "y",
+  "z"
+];
+
+console.log("a1:", a1);
+
+var AllLabels = {
+  foo: foo$1,
+  ptl: ptl$1,
+  a1: a1
+};
+
+function foo$2(x, y, z, dOpt) {
+  var d = dOpt !== undefined ? dOpt : "d=0";
+  return [
+          x,
+          y,
+          z,
+          d
+        ];
+}
+
+function ptl$2(none, extra, extra$1) {
+  return foo$2(none, "y", extra, extra$1);
+}
+
+var b1 = ptl$2("x", "z", undefined);
+
+console.log("b1:", b1);
+
+var b2 = ptl$2("x", "z", "d<-100");
+
+console.log("b2:", b2);
+
+var OptAtEnd = {
+  foo: foo$2,
+  ptl: ptl$2,
+  b1: b1,
+  b2: b2
+};
+
+function foo$3(d1Opt, x, d2Opt, y, d3Opt, z, d4Opt, w, d5Opt) {
+  var d1 = d1Opt !== undefined ? d1Opt : "d1=0";
+  var d2 = d2Opt !== undefined ? d2Opt : "d2=0";
+  var d3 = d3Opt !== undefined ? d3Opt : "d3=0";
+  var d4 = d4Opt !== undefined ? d4Opt : "d4=0";
+  var d5 = d5Opt !== undefined ? d5Opt : "d5=0";
+  return [
+          d1,
+          x,
+          d2,
+          y,
+          d3,
+          z,
+          d4,
+          w,
+          d5
+        ];
+}
+
+function ptl$3(none, none$1, none$2, none$3, none$4, none$5, extra) {
+  return foo$3(none, none$1, none$2, "y", none$3, none$4, none$5, "w", extra);
+}
+
+var c1 = ptl$3(undefined, "x", undefined, undefined, "z", undefined, undefined);
+
+console.log("c1:", c1);
+
+var c2 = ptl$3("d1<-100", "x", undefined, undefined, "z", undefined, undefined);
+
+console.log("c2:", c2);
+
+var c3 = ptl$3(undefined, "x", "d2<-200", undefined, "z", "d4<-400", undefined);
+
+console.log("c3:", c3);
+
+var OptMixed = {
+  foo: foo$3,
+  ptl: ptl$3,
+  c1: c1,
+  c2: c2,
+  c3: c3
+};
+
 exports.foo = foo;
 exports.z = z;
 exports.bar = bar;
@@ -70,4 +172,7 @@ exports.bar3 = bar3;
 exports.q = q;
 exports.inl = inl;
 exports.inl2 = inl2;
+exports.AllLabels = AllLabels;
+exports.OptAtEnd = OptAtEnd;
+exports.OptMixed = OptMixed;
 /*  Not a pure module */